+++ atlasdb-impl-shared/src/test/java/com/palantir/atlasdb/sweep/queue/TargetedSweeperTest.java
@@ -44,6 +44,7 @@ import org.mockito.ArgumentCaptor;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Iterables;
+import com.google.common.collect.Multimap;
 import com.palantir.atlasdb.cleaner.KeyValueServicePuncherStore;
 import com.palantir.atlasdb.cleaner.PuncherStore;
 import com.palantir.atlasdb.encoding.PtBytes;
@@ -115,23 +116,23 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
 
     @Test
     public void sweepStrategyNothingDoesNotPersistAnything() {
-        enqueueWrite(TABLE_NOTH, LOW_TS);
-        enqueueWrite(TABLE_NOTH, LOW_TS2);
+        enqueueWriteCommitted(TABLE_NOTH, LOW_TS);
+        enqueueWriteCommitted(TABLE_NOTH, LOW_TS2);
         verify(spiedKvs, times(2)).put(eq(TABLE_NOTH), anyMap(), anyLong());
         verify(spiedKvs, times(2)).put(any(TableReference.class), anyMap(), anyLong());
     }
 
     @Test
     public void sweepStrategyNothingDoesNotUpdateMetrics() {
-        enqueueWrite(TABLE_NOTH, LOW_TS);
-        enqueueWrite(TABLE_NOTH, LOW_TS2);
+        enqueueWriteCommitted(TABLE_NOTH, LOW_TS);
+        enqueueWriteCommitted(TABLE_NOTH, LOW_TS2);
         TargetedSweepMetricsTest.assertEnqueuedWritesConservativeEquals(0);
         TargetedSweepMetricsTest.assertEnqueuedWritesThoroughEquals(0);
     }
 
     @Test
     public void conservativeSweepAddsSentinelAndLeavesSingleValue() {
-        enqueueWrite(TABLE_CONS, LOW_TS);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS);
         assertReadAtTimestampReturnsNothing(TABLE_CONS, LOW_TS);
 
         sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
@@ -141,7 +142,7 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
 
     @Test
     public void sweepWithSingleEntryUpdatesMetrics() {
-        enqueueWrite(TABLE_CONS, LOW_TS);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS);
         sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
 
         TargetedSweepMetricsTest.assertTombstonesPutConservativeEquals(1);
@@ -154,7 +155,7 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
 
     @Test
     public void thoroughSweepDoesNotAddSentinelAndLeavesSingleValue() {
-        enqueueWrite(TABLE_THOR, LOW_TS);
+        enqueueWriteCommitted(TABLE_THOR, LOW_TS);
         assertReadAtTimestampReturnsNothing(TABLE_THOR, LOW_TS);
 
         sweepQueue.sweepNextBatch(ShardAndStrategy.thorough(THOR_SHARD));
@@ -166,15 +167,15 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
     public void thoroughSweepDeletesExistingSentinel() {
         spiedKvs.addGarbageCollectionSentinelValues(TABLE_THOR, ImmutableList.of(DEFAULT_CELL));
         assertReadAtTimestampReturnsSentinel(TABLE_THOR, 0L);
-        enqueueWrite(TABLE_THOR, 10L);
+        enqueueWriteCommitted(TABLE_THOR, 10L);
         sweepQueue.sweepNextBatch(ShardAndStrategy.thorough(THOR_SHARD));
         assertReadAtTimestampReturnsNothing(TABLE_THOR, 0L);
     }
 
     @Test
     public void conservativeSweepDeletesLowerValue() {
-        enqueueWrite(TABLE_CONS, LOW_TS);
-        enqueueWrite(TABLE_CONS, LOW_TS2);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS2);
         assertTestValueEnqueuedAtGivenTimestampStillPresent(TABLE_CONS, LOW_TS);
         assertTestValueEnqueuedAtGivenTimestampStillPresent(TABLE_CONS, LOW_TS2);
 
@@ -185,8 +186,8 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
 
     @Test
     public void thoroughSweepDeletesLowerValue() {
-        enqueueWrite(TABLE_THOR, LOW_TS);
-        enqueueWrite(TABLE_THOR, LOW_TS2);
+        enqueueWriteCommitted(TABLE_THOR, LOW_TS);
+        enqueueWriteCommitted(TABLE_THOR, LOW_TS2);
         assertTestValueEnqueuedAtGivenTimestampStillPresent(TABLE_THOR, LOW_TS);
         assertTestValueEnqueuedAtGivenTimestampStillPresent(TABLE_THOR, LOW_TS2);
 
@@ -197,8 +198,8 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
 
     @Test
     public void conservativeSweepCallsFollower() {
-        enqueueWrite(TABLE_CONS, LOW_TS);
-        enqueueWrite(TABLE_CONS, LOW_TS2);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS2);
         sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
 
         ArgumentCaptor<Set> captor = ArgumentCaptor.forClass(Set.class);
@@ -208,8 +209,8 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
 
     @Test
     public void thoroughSweepCallsFollower() {
-        enqueueWrite(TABLE_THOR, LOW_TS);
-        enqueueWrite(TABLE_THOR, LOW_TS2);
+        enqueueWriteCommitted(TABLE_THOR, LOW_TS);
+        enqueueWriteCommitted(TABLE_THOR, LOW_TS2);
         sweepQueue.sweepNextBatch(ShardAndStrategy.thorough(THOR_SHARD));
 
         ArgumentCaptor<Set> captor = ArgumentCaptor.forClass(Set.class);
@@ -221,7 +222,7 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
     public void conservativeSweepDeletesAllButLatestWithSingleDeleteAllTimestamps() {
         long lastWriteTs = TS_FINE_GRANULARITY - 1;
         for (long i = 0; i <= lastWriteTs; i++) {
-            enqueueWrite(TABLE_CONS, i);
+            enqueueWriteCommitted(TABLE_CONS, i);
         }
         sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
         assertReadAtTimestampReturnsSentinel(TABLE_CONS, lastWriteTs);
@@ -233,7 +234,7 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
     public void thoroughSweepDeletesAllButLatestWithSingleDeleteAllTimestampsIncludingSentinels() {
         long lastWriteTs = TS_FINE_GRANULARITY - 1;
         for (long i = 0; i <= lastWriteTs; i++) {
-            enqueueWrite(TABLE_THOR, i);
+            enqueueWriteCommitted(TABLE_THOR, i);
         }
         sweepQueue.sweepNextBatch(ShardAndStrategy.thorough(THOR_SHARD));
         assertReadAtTimestampReturnsNothing(TABLE_THOR, lastWriteTs);
@@ -243,9 +244,9 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
 
     @Test
     public void onlySweepsOneBatchAtATime() {
-        enqueueWrite(TABLE_CONS, LOW_TS);
-        enqueueWrite(TABLE_CONS, LOW_TS2);
-        enqueueWrite(TABLE_CONS, TS_FINE_GRANULARITY);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS2);
+        enqueueWriteCommitted(TABLE_CONS, TS_FINE_GRANULARITY);
 
         sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
         assertReadAtTimestampReturnsSentinel(TABLE_CONS, LOW_TS + 1);
@@ -255,7 +256,7 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
 
     @Test
     public void sweepDeletesWritesWhenTombstoneHasHigherTimestamp() {
-        enqueueWrite(TABLE_CONS, LOW_TS);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS);
         enqueueTombstone(TABLE_CONS, LOW_TS2);
         assertTestValueEnqueuedAtGivenTimestampStillPresent(TABLE_CONS, LOW_TS);
         assertReadAtTimestampReturnsTombstoneAtTimestamp(TABLE_CONS, LOW_TS2 + 1, LOW_TS2);
@@ -280,7 +281,7 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
     @Test
     public void sweepDeletesTombstonesWhenWriteHasHigherTimestamp() {
         enqueueTombstone(TABLE_CONS, LOW_TS);
-        enqueueWrite(TABLE_CONS, LOW_TS2);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS2);
         assertReadAtTimestampReturnsTombstoneAtTimestamp(TABLE_CONS, LOW_TS + 1, LOW_TS);
         assertTestValueEnqueuedAtGivenTimestampStillPresent(TABLE_CONS, LOW_TS2);
 
@@ -291,11 +292,11 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
 
     @Test
     public void sweepHandlesSequencesOfDeletesAndReadditionsInOneShot() {
-        enqueueWrite(TABLE_CONS, LOW_TS);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS);
         enqueueTombstone(TABLE_CONS, LOW_TS + 2);
-        enqueueWrite(TABLE_CONS, LOW_TS + 4);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS + 4);
         enqueueTombstone(TABLE_CONS, LOW_TS + 6);
-        enqueueWrite(TABLE_CONS, LOW_TS + 8);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS + 8);
 
         sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
         for (int i = 0; i < 10; i = i + 2) {
@@ -315,10 +316,10 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
     public void sweepProgressesAndSkipsEmptyFinePartitions() {
         long tsFineTwo = LOW_TS + TS_FINE_GRANULARITY;
         long tsFineFour = LOW_TS + 3 * TS_FINE_GRANULARITY;
-        enqueueWrite(TABLE_CONS, LOW_TS);
-        enqueueWrite(TABLE_CONS, tsFineTwo);
-        enqueueWrite(TABLE_CONS, tsFineFour);
-        enqueueWrite(TABLE_CONS, tsFineFour + 1L);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS);
+        enqueueWriteCommitted(TABLE_CONS, tsFineTwo);
+        enqueueWriteCommitted(TABLE_CONS, tsFineFour);
+        enqueueWriteCommitted(TABLE_CONS, tsFineFour + 1L);
 
         // first sweep effectively only writes a sentinel
         sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
@@ -348,10 +349,10 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
     public void sweepProgressesAcrossCoarsePartitions() {
         long tsCoarseTwo = LOW_TS + TS_FINE_GRANULARITY + TS_COARSE_GRANULARITY;
         long tsCoarseFour = LOW_TS + 3 * TS_COARSE_GRANULARITY;
-        enqueueWrite(TABLE_CONS, LOW_TS);
-        enqueueWrite(TABLE_CONS, tsCoarseTwo);
-        enqueueWrite(TABLE_CONS, tsCoarseFour);
-        enqueueWrite(TABLE_CONS, tsCoarseFour + 1L);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS);
+        enqueueWriteCommitted(TABLE_CONS, tsCoarseTwo);
+        enqueueWriteCommitted(TABLE_CONS, tsCoarseFour);
+        enqueueWriteCommitted(TABLE_CONS, tsCoarseFour + 1L);
 
         sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
         assertReadAtTimestampReturnsSentinel(TABLE_CONS, LOW_TS);
@@ -375,8 +376,8 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
     @Test
     public void sweepProgressesToEndOfPartitionWhenFewValuesAndSweepTsLarge() {
         long writeTs = getSweepTsCons() - 3 * TS_FINE_GRANULARITY;
-        enqueueWrite(TABLE_CONS, writeTs);
-        enqueueWrite(TABLE_CONS, writeTs + 5);
+        enqueueWriteCommitted(TABLE_CONS, writeTs);
+        enqueueWriteCommitted(TABLE_CONS, writeTs + 5);
 
         sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
         assertProgressUpdatedToTimestamp(maxTsForFinePartition(tsPartitionFine(writeTs)));
@@ -390,7 +391,7 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
         immutableTs = 2 * TS_COARSE_GRANULARITY - TS_FINE_GRANULARITY;
         verify(spiedKvs, never()).deleteAllTimestamps(any(TableReference.class), anyMap(), eq(false));
 
-        enqueueWrite(TABLE_CONS, immutableTs - 1);
+        enqueueWriteCommitted(TABLE_CONS, immutableTs - 1);
         sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
         verify(spiedKvs, times(1)).deleteAllTimestamps(any(TableReference.class), anyMap(), eq(false));
 
@@ -402,9 +403,9 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
     public void sweepableTimestampsGetsScrubbedWhenNoMoreToSweepButSweepTsInNewCoarsePartition() {
         long tsSecondPartitionFine = LOW_TS + TS_FINE_GRANULARITY;
         long largestFirstPartitionCoarse = TS_COARSE_GRANULARITY - 1L;
-        enqueueWrite(TABLE_CONS, LOW_TS);
-        enqueueWrite(TABLE_CONS, tsSecondPartitionFine);
-        enqueueWrite(TABLE_CONS, largestFirstPartitionCoarse);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS);
+        enqueueWriteCommitted(TABLE_CONS, tsSecondPartitionFine);
+        enqueueWriteCommitted(TABLE_CONS, largestFirstPartitionCoarse);
 
         sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
         assertLowestFinePartitionInSweepableTimestampsEquals(tsPartitionFine(LOW_TS));
@@ -421,10 +422,10 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
         long tsSecondPartitionFine = LOW_TS + TS_FINE_GRANULARITY;
         long largestFirstPartitionCoarse = TS_COARSE_GRANULARITY - 1L;
         long thirdPartitionCoarse = 2 * TS_COARSE_GRANULARITY;
-        enqueueWrite(TABLE_CONS, LOW_TS);
-        enqueueWrite(TABLE_CONS, tsSecondPartitionFine);
-        enqueueWrite(TABLE_CONS, largestFirstPartitionCoarse);
-        enqueueWrite(TABLE_CONS, thirdPartitionCoarse);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS);
+        enqueueWriteCommitted(TABLE_CONS, tsSecondPartitionFine);
+        enqueueWriteCommitted(TABLE_CONS, largestFirstPartitionCoarse);
+        enqueueWriteCommitted(TABLE_CONS, thirdPartitionCoarse);
 
         sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
         assertLowestFinePartitionInSweepableTimestampsEquals(tsPartitionFine(LOW_TS));
@@ -439,10 +440,10 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
     @Test
     public void sweepableCellsGetsScrubbedWheneverLastSweptInNewPartition() {
         long tsSecondPartitionFine = LOW_TS + TS_FINE_GRANULARITY;
-        enqueueWrite(TABLE_CONS, LOW_TS);
-        enqueueWrite(TABLE_CONS, LOW_TS + 1L);
-        enqueueWrite(TABLE_CONS, tsSecondPartitionFine);
-        enqueueWrite(TABLE_CONS, getSweepTsCons());
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS + 1L);
+        enqueueWriteCommitted(TABLE_CONS, tsSecondPartitionFine);
+        enqueueWriteCommitted(TABLE_CONS, getSweepTsCons());
 
         // last swept timestamp: TS_FINE_GRANULARITY - 1
         sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
@@ -471,9 +472,9 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
     @Test
     public void doesNotTransitivelyRetainWritesFromBeforeSweepTimestamp() {
         long sweepTimestamp = getSweepTsCons();
-        enqueueWrite(TABLE_CONS, sweepTimestamp - 10);
+        enqueueWriteCommitted(TABLE_CONS, sweepTimestamp - 10);
         enqueueTombstone(TABLE_CONS, sweepTimestamp - 5);
-        enqueueWrite(TABLE_CONS, sweepTimestamp + 5);
+        enqueueWriteCommitted(TABLE_CONS, sweepTimestamp + 5);
 
         sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
         assertReadAtTimestampReturnsSentinel(TABLE_CONS, sweepTimestamp - 5);
@@ -512,9 +513,9 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
         long newPartitionFirstTs = baseSweepTs + 5;
         long newPartitionSecondTs = baseSweepTs + 10;
 
-        enqueueWrite(TABLE_CONS, oldPartitionTs);
-        enqueueWrite(TABLE_CONS, newPartitionFirstTs);
-        enqueueWrite(TABLE_CONS, newPartitionSecondTs);
+        enqueueWriteCommitted(TABLE_CONS, oldPartitionTs);
+        enqueueWriteCommitted(TABLE_CONS, newPartitionFirstTs);
+        enqueueWriteCommitted(TABLE_CONS, newPartitionSecondTs);
 
         assertReadAtTimestampReturnsNothing(TABLE_CONS, oldPartitionTs);
 
@@ -531,9 +532,9 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
 
     @Test
     public void doesNotGoBackwardsEvenIfSweepTimestampRegressesWithinBucket() {
-        enqueueWrite(TABLE_CONS, LOW_TS);
-        enqueueWrite(TABLE_CONS, LOW_TS2);
-        enqueueWrite(TABLE_CONS, LOW_TS3);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS2);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS3);
 
         runConservativeSweepAtTimestamp(LOW_TS2 + 5);
         assertReadAtTimestampReturnsSentinel(TABLE_CONS, LOW_TS2);
@@ -554,9 +555,9 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
 
     @Test
     public void canSweepAtMinimumTime() {
-        enqueueWrite(TABLE_CONS, LOW_TS);
-        enqueueWrite(TABLE_CONS, LOW_TS2);
-        enqueueWrite(TABLE_CONS, LOW_TS3);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS2);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS3);
 
         runConservativeSweepAtTimestamp(Long.MIN_VALUE);
         assertTestValueEnqueuedAtGivenTimestampStillPresent(TABLE_CONS, LOW_TS);
@@ -565,9 +566,9 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
 
     @Test
     public void canSweepAtMaximumTime() {
-        enqueueWrite(TABLE_CONS, LOW_TS);
-        enqueueWrite(TABLE_CONS, LOW_TS2);
-        enqueueWrite(TABLE_CONS, LOW_TS3);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS2);
+        enqueueWriteCommitted(TABLE_CONS, LOW_TS3);
 
         runConservativeSweepAtTimestamp(Long.MAX_VALUE);
         assertReadAtTimestampReturnsSentinel(TABLE_CONS, LOW_TS3);
@@ -577,9 +578,9 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
     @Test
     public void doesNotGoBackwardsEvenIfSweepTimestampRegressesAcrossBoundary() {
         long coarseBoundary = TS_COARSE_GRANULARITY;
-        enqueueWrite(TABLE_CONS, coarseBoundary - 5);
-        enqueueWrite(TABLE_CONS, coarseBoundary + 5);
-        enqueueWrite(TABLE_CONS, coarseBoundary + 15);
+        enqueueWriteCommitted(TABLE_CONS, coarseBoundary - 5);
+        enqueueWriteCommitted(TABLE_CONS, coarseBoundary + 5);
+        enqueueWriteCommitted(TABLE_CONS, coarseBoundary + 15);
 
         // Need 2 sweeps to get through the first coarse bucket
         runConservativeSweepAtTimestamp(coarseBoundary + 8);
@@ -619,10 +620,72 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
         TargetedSweepMetricsTest.assertSweepTimestampThoroughEquals(5);
     }
 
+    @Test
+    public void doNotSweepAnythingAfterEntryWithCommitTsAfterSweepTs() {
+        immutableTs = 1000L;
+        // put 4 writes committed at the same timestamp as start timestamp, and put one committed at sweep timestamp
+        enqueueWriteCommitted(TABLE_CONS, 900);
+        enqueueWriteCommitted(TABLE_CONS, 910);
+        enqueueWriteCommitted(TABLE_CONS, 920);
+        enqueueWriteCommitedAt(TABLE_CONS, 950, immutableTs);
+        enqueueWriteCommitted(TABLE_CONS, 970);
+
+        sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
+        assertThat(progress.getLastSweptTimestamp(ShardAndStrategy.conservative(CONS_SHARD))).isEqualTo(920L);
+        ArgumentCaptor<Map> argument = ArgumentCaptor.forClass(Map.class);
+        verify(spiedKvs, times(1)).deleteAllTimestamps(eq(TABLE_CONS), argument.capture(), eq(false));
+        assertThat(argument.getValue()).containsValue(920L);
+
+        sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
+        assertThat(progress.getLastSweptTimestamp(ShardAndStrategy.conservative(CONS_SHARD))).isEqualTo(920L);
+        verify(spiedKvs, times(1)).deleteAllTimestamps(any(TableReference.class), anyMap(), eq(false));
+
+        immutableTs = 1001L;
+        sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
+        assertThat(progress.getLastSweptTimestamp(ShardAndStrategy.conservative(CONS_SHARD))).isEqualTo(1001L - 1L);
+        // we have now had a total of 2 calls to deleteAllTimestamps, 1 from before and one new
+        verify(spiedKvs, times(2)).deleteAllTimestamps(eq(TABLE_CONS), argument.capture(), eq(false));
+        assertThat(argument.getValue()).containsValue(970L);
+    }
+
+    @Test
+    public void doNotDeleteAnythingAfterEntryWithCommitTsAfterSweepTs() {
+        immutableTs = 1000L;
+        enqueueWriteUncommitted(TABLE_CONS, 900);
+        enqueueWriteUncommitted(TABLE_CONS, 920);
+        enqueueWriteCommitedAt(TABLE_CONS, 950, 2000);
+        enqueueWriteUncommitted(TABLE_CONS, 970);
+        enqueueWriteUncommitted(TABLE_CONS, 1110);
+
+        sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
+        assertThat(progress.getLastSweptTimestamp(ShardAndStrategy.conservative(CONS_SHARD))).isEqualTo(920L);
+        verify(spiedKvs, never()).deleteAllTimestamps(any(TableReference.class), anyMap(), eq(false));
+
+        ArgumentCaptor<Multimap> multimap = ArgumentCaptor.forClass(Multimap.class);
+        verify(spiedKvs, times(1)).delete(eq(TABLE_CONS), multimap.capture());
+        assertThat(multimap.getValue().keySet()).containsExactly(DEFAULT_CELL);
+        assertThat(multimap.getValue().values()).containsExactly(900L, 920L);
+
+        sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
+        assertThat(progress.getLastSweptTimestamp(ShardAndStrategy.conservative(CONS_SHARD))).isEqualTo(920L);
+        verify(spiedKvs, never()).deleteAllTimestamps(any(TableReference.class), anyMap(), eq(false));
+        verify(spiedKvs, times(1)).delete(any(TableReference.class), any(Multimap.class));
+        assertReadAtTimestampReturnsValue(TABLE_CONS, 1500L, 1110L);
+
+        immutableTs = 2009L;
+        sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD));
+        assertThat(progress.getLastSweptTimestamp(ShardAndStrategy.conservative(CONS_SHARD))).isEqualTo(2009L - 1L);
+        ArgumentCaptor<Map> map = ArgumentCaptor.forClass(Map.class);
+        verify(spiedKvs, times(1)).deleteAllTimestamps(eq(TABLE_CONS), map.capture(), eq(false));
+        assertThat(map.getValue()).containsValue(950L);
+
+        assertReadAtTimestampReturnsValue(TABLE_CONS, 1500L, 950L);
+    }
+
     private void writeValuesAroundSweepTimestampAndSweepAndCheck(long sweepTimestamp, int sweepIterations) {
-        enqueueWrite(TABLE_CONS, sweepTimestamp - 10);
-        enqueueWrite(TABLE_CONS, sweepTimestamp - 5);
-        enqueueWrite(TABLE_CONS, sweepTimestamp + 5);
+        enqueueWriteCommitted(TABLE_CONS, sweepTimestamp - 10);
+        enqueueWriteCommitted(TABLE_CONS, sweepTimestamp - 5);
+        enqueueWriteCommitted(TABLE_CONS, sweepTimestamp + 5);
 
         IntStream.range(0, sweepIterations)
                 .forEach(unused -> sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(CONS_SHARD)));
@@ -637,14 +700,23 @@ public class TargetedSweeperTest extends AbstractSweepQueueTest {
                 desiredSweepTimestamp);
     }
 
-    private void enqueueWrite(TableReference tableRef, long ts) {
+    private void enqueueWriteCommitted(TableReference tableRef, long ts) {
         putTimestampIntoTransactionTable(ts, ts);
-        sweepQueue.enqueue(writeToDefaultCell(tableRef, ts), ts);
+        enqueueWriteUncommitted(tableRef, ts);
+    }
+
+    private void enqueueWriteCommitedAt(TableReference tableRef, long startTs, long commitTs) {
+        putTimestampIntoTransactionTable(startTs, commitTs);
+        enqueueWriteUncommitted(tableRef, startTs);
     }
 
     private void enqueueTombstone(TableReference tableRef, long ts) {
         putTimestampIntoTransactionTable(ts, ts);
         sweepQueue.enqueue(tombstoneToDefaultCell(tableRef, ts), ts);
+    }
+
+    private void enqueueWriteUncommitted(TableReference tableRef, long startTs) {
+        sweepQueue.enqueue(writeToDefaultCell(tableRef, startTs), startTs);
     }
 
     private Map<TableReference, ? extends Map<Cell, byte[]>> writeToDefaultCell(TableReference tableRef, long ts) {